Fix integration tests
