+++ nvflare/private/fed/app/deployer/simulator_deployer.py
@@ -17,7 +17,7 @@ import tempfile
 
 from nvflare.apis.event_type import EventType
 from nvflare.fuel.utils.network_utils import get_open_ports
-from nvflare.private.fed.app.server.server_train import create_admin_server
+from nvflare.private.fed.app.utils import create_admin_server
 from nvflare.private.fed.client.admin import FedAdminAgent
 from nvflare.private.fed.client.admin_msg_sender import AdminMessageSender
 from nvflare.private.fed.client.client_req_processors import ClientRequestProcessors
