+++ nvflare/private/fed/app/simulator/simulator_worker.py
@@ -26,7 +26,7 @@ from nvflare.apis.fl_constant import FLContextKey, WorkspaceConstants
 from nvflare.fuel.common.multi_process_executor_constants import CommunicationMetaData
 from nvflare.fuel.hci.server.authz import AuthorizationService
 from nvflare.fuel.sec.audit import AuditService
-from nvflare.private.fed.app.client.worker_process import check_parent_alive
+from nvflare.private.fed.app.utils import check_parent_alive
 from nvflare.private.fed.client.admin import FedAdminAgent
 from nvflare.private.fed.client.admin_msg_sender import AdminMessageSender
 from nvflare.private.fed.client.client_req_processors import ClientRequestProcessors
