+++ tests/integration_test/src/admin_controller.py
@@ -37,7 +37,7 @@ from tests.integration_test.src.action_handlers import (
     _TestDoneHandler,
 )
 from tests.integration_test.src.site_launcher import SiteLauncher
-from tests.integration_test.src.utils import get_job_run_statuses
+from tests.integration_test.src.utils import get_job_meta
 
 
 def _parse_workflow_states(stats_message: dict):
@@ -356,13 +356,14 @@ class AdminController:
 
         self.test_done = False
         while not self.test_done:
-            job_run_statuses = get_job_run_statuses(self.super_admin_api)
+            job_meta = get_job_meta(self.super_admin_api, job_id=self.job_id)
+            job_run_status = job_meta.get("status")
 
             if self.job_id:
                 stats = self._get_stats(target=TargetType.SERVER, job_id=self.job_id)
                 # update run_state
                 changed, run_state = self._update_run_state(
-                    stats=stats, run_state=run_state, job_run_status=job_run_statuses.get(self.job_id, "")
+                    stats=stats, run_state=run_state, job_run_status=job_run_status
                 )
 
             if event_idx < len(event_sequence):
